pyproject.toml
[tool.poetry]
name = "crawler framework"
version = "0.1.0"
description = "Write an application to crawl an online news website and store the data in BigQuery"
authors = ["Debora Ratto <drattodourado@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.1"
scrapy = "^2.11.1"
scrapy-splash = "^0.9.0"
google-cloud-bigquery = "^3.19.0"
pandas = "^2.2.1"
pandas-gbq = "^0.22.0"
flask = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
